Introduce MemberIdent/LoanIdent newtypes and widen IDs
Replace String idents with MemberIdent/LoanIdent newtypes, update repo
interfaces and HTTP handlers to use them, and change several numeric ID
types to i32 in domain and persistence with corresponding conversions

+++ server/crates/application/src/commands/lending.rs
@@ -43,7 +43,7 @@ impl LendingCommands {
             .await
             .context("Failed to count active loans for member")?;
         anyhow::ensure!(
-            member.can_check_out_more_books(active_loan_count),
+            member.can_check_out_more_books(active_loan_count as i16),
             MemberError::LoanLimitReached
         );
